+++ public/app/main-page/main-page.html
@@ -1,4 +1,4 @@
-<div ng-controller="dashBoardCtrl" layout="column" style="height:500px;" ng-cloak>
+<div ng-controller="mainPageCtrl" layout="column" style="height:500px;" ng-cloak>
 
   <section layout="row" flex>
 
@@ -38,9 +38,10 @@
             <md-icon class="link" ng-click="toggleLeft()" hide-gt-sm style="margin: 16px;height: 30px;width: 30px;" md-svg-icon="/assets/images/icons/menu.svg"></md-icon>
             <md-icon id="menu" class="link" ng-click="toggle()"  style="margin: 16px;height: 30px;width: 30px;" md-svg-icon="/assets/images/icons/menu.svg"></md-icon>
             <span flex></span>
-            <div id="profile" class="link-background link" ng-click="showToolbarMenu($event)" style="margin-right: 15px;padding: 22px;font-size: 14px;">
+            <div id="profile" class="link-background link " ng-click="showToolbarMenu($event)" style="margin-right: 15px;padding: 22px;font-size: 14px;">
             <i class="icon-user" ></i>
             <strong >{{username}}</strong>
+            {{test}}
             <i class="icon-caret-down" ></i>
             </div>    
         </md-toolbar>
