--- src/app/dashboard/dashboard-module.js
@@ -1,5 +1,0 @@
-"use strict";
-require('angular-material');
-var dashBoardCtrl = require('./dashboard-controller.js');
-var dashBoardModule = angular.module('myApp.dashboard',['ngRoute','ngMaterial']);
-dashBoardModule.controller('dashBoardCtrl',dashBoardCtrl);
