+++ src/app/main-page/main-page-module.js
@@ -1,0 +1,5 @@
+"use strict";
+require('angular-material');
+var mainPageCtrl = require('./main-page-controller.js');
+var dashBoardModule = angular.module('myApp.mainPage',['ngRoute','ngMaterial']);
+dashBoardModule.controller('mainPageCtrl',mainPageCtrl);
