+++ src/app/main-page/main-page.html
@@ -1,4 +1,4 @@
-<div ng-controller="dashBoardCtrl" layout="column" style="height:500px;" ng-cloak>
+<div ng-controller="mainPageCtrl" layout="column" style="height:500px;" ng-cloak>
 
   <section layout="row" flex>
